Working optimizer based on splitting the top 10
Optimize the BLOSUM50 matrix by updating each entry based on the
log odds of a pair occuring in the lowest scoring positive
examples minus the log odds of a pair occuring in the highest
scoring negative examples. This should drive the two distributions
apart, leading to reduced false positive rates.